Starting point for a DNS server answering lookups in response to DynDNS.
When a Fritz!Box registers via DynDNS, it tries to verify whether the
registration of its host name was successful by looking up its host name
via DNS. If this lookup fails, the box post error messages to its log
and re-trys to register multiple times. To prevent this unnecessary
traffic, a DNS server is required that responds to the boxes requests.