Issue #91: Keep track of blocks of phone number reported as SPAM.
When a user wants to detect ranges of phone numbers that obviously
belong to the same SPAM source, there must be an easy way to decide
whether a certain number belongs to such block. To accomplish this, two
additional database tables are maintained with aggregation information.
SPAMREPORTS_10 and SPAMREPORTS_100 with aggregate information of blocks
of size 10 and blocks of size 100 respectively.